ì 아ë이디찾기까지완료 정보수정에서 ë비번은 안됨

+++ src/main/java/generationgap/co/kr/domain/chat/ChatMessage.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 public class ChatMessage {
     private Long messagesIdx;
     private Long groupChatIdx;
-    private Integer senderIdx;
+    private Long senderIdx;
     private String nickname;
     private String content;
     private LocalDateTime sentAt;
